+++ packages/ethers-sdk/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.9.2](https://github.com/bosonprotocol/core-components/compare/@bosonprotocol/ethers-sdk@1.9.1...@bosonprotocol/ethers-sdk@1.9.2) (2022-10-13)
+
+**Note:** Version bump only for package @bosonprotocol/ethers-sdk
+
+
+
+
+
+## [1.9.1](https://github.com/bosonprotocol/core-components/compare/@bosonprotocol/ethers-sdk@1.9.0...@bosonprotocol/ethers-sdk@1.9.1) (2022-10-13)
+
+**Note:** Version bump only for package @bosonprotocol/ethers-sdk
+
+
+
+
+
 # [1.9.0](https://github.com/bosonprotocol/core-components/compare/@bosonprotocol/ethers-sdk@1.8.0...@bosonprotocol/ethers-sdk@1.9.0) (2022-10-12)