+++ packages/react-kit/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.9.2](https://github.com/bosonprotocol/core-components/compare/@bosonprotocol/react-kit@0.9.1...@bosonprotocol/react-kit@0.9.2) (2022-10-13)
+
+**Note:** Version bump only for package @bosonprotocol/react-kit
+
+
+
+
+
+## [0.9.1](https://github.com/bosonprotocol/core-components/compare/@bosonprotocol/react-kit@0.9.0...@bosonprotocol/react-kit@0.9.1) (2022-10-13)
+
+**Note:** Version bump only for package @bosonprotocol/react-kit
+
+
+
+
+
 # [0.9.0](https://github.com/bosonprotocol/core-components/compare/@bosonprotocol/react-kit@0.8.0...@bosonprotocol/react-kit@0.9.0) (2022-10-12)